Refactored liveinfo packages /app/src/app/pages/Sandbox/Editor/… (#2878)
